pyproject.toml
[project]
name = "document-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=1.4.1",
    "ipynb>=0.5.1",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "openai>=2.18.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "transformers>=5.1.0",
    "wikipedia>=1.4.0",
]
